+++ GEN_QUIDVUE/FLAT_VUE/CSGenio.core/business/Arrays/ArrayAperiodi.cs
@@ -5,7 +5,7 @@ namespace CSGenio.business
 	/// <summary>
 	/// Array aperiodi (Periodicidade de Recolha)
 	/// </summary>
-	public class ArrayAperiodi : Array<double>
+	public class ArrayAperiodi : Array<decimal>
 	{
 		/// <summary>
 		/// The instance
@@ -28,47 +28,47 @@ namespace CSGenio.business
 		/// <summary>
 		/// Horário
 		/// </summary>
-		public const double E_1_1 = 1;
+		public const decimal E_1_1 = 1M;
 		/// <summary>
 		/// Diário
 		/// </summary>
-		public const double E_2_2 = 2;
+		public const decimal E_2_2 = 2M;
 		/// <summary>
 		/// Semanal
 		/// </summary>
-		public const double E_3_3 = 3;
+		public const decimal E_3_3 = 3M;
 		/// <summary>
 		/// Mensal
 		/// </summary>
-		public const double E_4_4 = 4;
+		public const decimal E_4_4 = 4M;
 		/// <summary>
 		/// Bimestral
 		/// </summary>
-		public const double E_10_5 = 10;
+		public const decimal E_10_5 = 10M;
 		/// <summary>
 		/// Trimestral
 		/// </summary>
-		public const double E_5_6 = 5;
+		public const decimal E_5_6 = 5M;
 		/// <summary>
 		/// Semestral
 		/// </summary>
-		public const double E_6_7 = 6;
+		public const decimal E_6_7 = 6M;
 		/// <summary>
 		/// Anual
 		/// </summary>
-		public const double E_7_8 = 7;
+		public const decimal E_7_8 = 7M;
 		/// <summary>
 		/// Variável
 		/// </summary>
-		public const double E_9_9 = 9;
+		public const decimal E_9_9 = 9M;
 		/// <summary>
 		/// Bianual
 		/// </summary>
-		public const double E_11_10 = 11;
+		public const decimal E_11_10 = 11M;
 		/// <summary>
 		/// 5 anos
 		/// </summary>
-		public const double E_12_11 = 12;
+		public const decimal E_12_11 = 12M;
 
 		/// <summary>
 		/// Prevents a default instance of the <see cref="ArrayAperiodi"/> class from being created.
@@ -79,9 +79,9 @@ namespace CSGenio.business
         /// Loads the dictionary.
         /// </summary>
         /// <returns></returns>
-		protected override Dictionary<double, ArrayElement> LoadDictionary()
+		protected override Dictionary<decimal, ArrayElement> LoadDictionary()
 		{
-			return new Dictionary<double, ArrayElement>()
+			return new Dictionary<decimal, ArrayElement>()
 			{
 				{ E_1_1, new ArrayElement() { ResourceId = "HORARIO56549", HelpId = "", Group = "" } },
 				{ E_2_2, new ArrayElement() { ResourceId = "DIARIO16236", HelpId = "", Group = "" } },
@@ -102,7 +102,7 @@ namespace CSGenio.business
 		/// </summary>
 		/// <param name="cod">The cod.</param>
 		/// <returns></returns>
-		public static string CodToDescricao(double cod)
+		public static string CodToDescricao(decimal cod)
 		{
 			return Instance.CodToDescricaoImpl(cod);
 		}
@@ -111,7 +111,7 @@ namespace CSGenio.business
 		/// Gets the elements.
 		/// </summary>
 		/// <returns></returns>
-		public static List<double> GetElements()
+		public static List<decimal> GetElements()
 		{
 			return Instance.GetElementsImpl();
 		}
@@ -123,14 +123,14 @@ namespace CSGenio.business
 		/// <returns></returns>
 		public static ArrayElement GetElement(string cod)
 		{
-            return Instance.GetElementImpl(double.Parse(cod));
+            return Instance.GetElementImpl(decimal.Parse(cod));
         }
 
 		/// <summary>
 		/// Gets the dictionary.
 		/// </summary>
 		/// <returns></returns>
-		public static IDictionary<double, string> GetDictionary()
+		public static IDictionary<decimal, string> GetDictionary()
 		{
 			return Instance.GetDictionaryImpl();
 		}
@@ -142,7 +142,7 @@ namespace CSGenio.business
 		/// <returns></returns>
 		public static string GetHelpId(string cod)
 		{
-			return Instance.GetHelpIdImpl(double.Parse(cod));
+			return Instance.GetHelpIdImpl(decimal.Parse(cod));
 		}
 	}
 }
